// app/dashboard/recruiter/jobs/new/types.ts
export interface JobData {
  jobTitle: string;
  companyName: string;
  jobType: string;
  experience: string;
  status: string;
  openings: string;
  jobLocation: string;
  logo?: string;

  companyOverview: string;
  companyLocation: string;
  jobDescription: string;
  skills: string;
  perks: string;
  salary: string;
  contactDetails: string;
  bondDetails: string;
}